Moved the different classes in the controller.cs into their own files.
