+++ LOLBITS/LOLBITS/Controlling/Content.cs
@@ -1,0 +1,10 @@
+﻿namespace LOLBITS
+{
+    public class Content
+    {
+        public string NextId { get; set; }
+        public string NextAuth { get; set; }
+        public string[] Commands { get; set; }
+
+    }
+}
