+++ LOLBITS/LOLBITS/Controlling/Response.cs
@@ -1,0 +1,13 @@
+﻿namespace LOLBITS
+{
+    public class Response
+    {   
+        public Response(string output, string reqId)
+        {
+            Output = output;
+            ReqId = reqId;
+        }
+        public string Output { get; set; }
+        public string ReqId { get; set; }
+    }
+}
